Moved mosaic code to image_utils.
